Renames references of "interval" to "segment"
I view an "interval" as a specific kind of workout segment where you
alternate between high and low powers. In the code, I was referring to
all kinds of workout segments as "interval"s, in more of the "time
interval" way. To resolve this confusion, I renamed the "interval" types
to "segment".

Fixes #5