Adds a "ramp" segment type which transitions between two powers
In order to make typescript better at testing my tagged unions, I
switched from if statements to switch statements for a bunch of places
where the type of a segment is checked.

I also changed how the status of the current segment is returned from
`findSegment`, because the current expected power is no longer the same
as the expected average power of the whole segment. This splits up
several of those values into separate properties so the page can
correctly interpret it.

Adding the ramps to the preferences segment display was complicated
because I wanted to accurately display the power zones within the chart.
I ended up figuring out how to do this by drawing a series of svg
trapezoids, one for each power zone.

Fixes #4